Retry play_url_media commands
I have observed the "Play" command at the end of play_url_media
return a 500 error (the error text was a uPnP 501 error with the
description "Action failed").  When this happens, a subsequent
"Play" command issued about 3.5 seconds later works.

To address this, let's retry the commands we send in the
play_url_media method.  I have not observed errors with the other
two ("Stop" and "SetAVTransportURI", but I don't currently see a
downside to retrying those as well).  This should make the
play_url_media method more robust, which should improve the
reliability when changing tracks.

Observed and tested with a Yamaha TSR-7810.